requirements.txt
speedtest-cli
python-twitter
schedule
python-dotenv